pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.10",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-ninja>=1.4.1",
    "django-ninja-extra>=0.30.0",
    "django-storages>=1.14.6",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.7",
    "whitenoise>=6.9.0",
]
